backend/nova_client.py
import boto3
import os
import json
from dotenv import load_dotenv

load_dotenv()

class NovaClient:
    """
    Centralized client for interacting with Amazon Bedrock Nova models.
    """
    def __init__(self):
        self.region = os.getenv("AWS_REGION", "us-east-1")
        self.bedrock_runtime = boto3.client(
            service_name="bedrock-runtime",
            region_name=self.region
        )
        self.nova_lite_id = os.getenv("NOVA_LITE_MODEL_ID", "amazon.nova-lite-v1:0")
        self.nova_sonic_id = os.getenv("NOVA_SONIC_MODEL_ID", "amazon.nova-sonic-v1:0")

    def invoke_nova_lite(self, prompt: str, system_prompt: str = ""):
        """
        Invokes Nova Lite for reasoning and text generation.
        """
        body = json.dumps({
            "inferenceConfig": {
                "maxTokens": 1000,
                "temperature": 0.5,
                "topP": 0.9
            },
            "messages": [
                {
                    "role": "user",
                    "content": [{"text": prompt}]
                }
            ],
            "system": [{"text": system_prompt}] if system_prompt else []
        })

        response = self.bedrock_runtime.invoke_model(
            modelId=self.nova_lite_id,
            body=body
        )
        
        response_body = json.loads(response.get("body").read())
        return response_body["output"]["message"]["content"][0]["text"]

    def discover_schemes(self, user_query: str):
        """
        Uses Nova Lite to analyze user needs and suggest government schemes.
        """
        try:
            prompt = (
                f"Analyze the following user situation and suggest relevant government schemes or subsidies they might qualify for: '{user_query}'. "
                "Return the response strictly in JSON format with a key 'eligible_schemes' which is a list of scheme names."
            )
            
            raw_response = self.invoke_nova_lite(prompt)
            data = json.loads(raw_response)
            return data
        except Exception as e:
            print(f"Error in discover_schemes: {e}")
            # Fallback for hackathon demo if model invocation fails (e.g. missing credentials)
            return {
                "eligible_schemes": [
                    "Electricity Subsidy",
                    "Energy Assistance Program",
                    "Low Income Housing Support"
                ]
            }

    def invoke_nova_vision(self, image_bytes: bytes, prompt: str):
        """
        Invokes Nova for image/document reasoning.
        """
        body = json.dumps({
            "inferenceConfig": {"maxTokens": 1000},
            "messages": [
                {
                    "role": "user",
                    "content": [
                        {"image": {"format": "png", "source": {"bytes": image_bytes.decode('latin1')}}},
                        {"text": prompt}
                    ]
                }
            ]
        })

        response = self.bedrock_runtime.invoke_model(
            modelId=self.nova_lite_id, # Nova Lite is multi-modal
            body=body
        )
        
        response_body = json.loads(response.get("body").read())
        return response_body["output"]["message"]["content"][0]["text"]
