backend/requirements.txt
fastapi
uvicorn
boto3
langchain
langchain-community
langchain-aws
pydantic
python-dotenv
python-multipart
